refactor: Move app initialization as well
